switch coefficients to short from int

+++ src/com/welty/novello/eval/CoefficientSet.java
@@ -8,20 +8,20 @@ package com.welty.novello.eval;
  * slices[nEmpty] is called a "slice"
 */
 class CoefficientSet {
-    private final int[][][] slices;
+    private final short[][][] slices;
     private final String name;
 
     public CoefficientSet(EvalStrategy strategy, String name) {
         this (readSlices(strategy, name), name);
     }
 
-    private CoefficientSet(int[][][] slices, String name) {
+    private CoefficientSet(short[][][] slices, String name) {
         this.slices = slices;
         this.name = name;
     }
 
-    private static int[][][] readSlices(EvalStrategy strategy, String name) {
-        int[][][] slices = new int[64][][];
+    private static short[][][] readSlices(EvalStrategy strategy, String name) {
+        short[][][] slices = new short[64][][];
         for (int nEmpty = 0; nEmpty < 64; nEmpty++) {
             slices[nEmpty] = strategy.readSlice(nEmpty, name);
         }
@@ -34,7 +34,7 @@ class CoefficientSet {
      * @param nEmpty number of empty squares
      * @return slice[iFeature][instance] is the eval coefficient
      */
-    public int[][] slice(int nEmpty) {
+    public short[][] slice(int nEmpty) {
         return slices[nEmpty];
     }
